Put service queries into guard
